combine bagger and creating folder (#18), update test
