--- araig_calculators/launch/test_rosbagger.launch
@@ -1,10 +1,0 @@
-<launch>        
-
-    <rosparam command="load" file="$(find araig_calculators)/config/test_rosbagger.yaml" />
-
-    <node pkg="araig_calculators" type="rosbagger_node" name="test_bagger" output="screen" >
-        <remap from="/start" to="/signal/runner/start_robot"/>
-        <remap from="/stop" to="/signal/runner/test_completed"/>
-    </node>
-
-</launch>
