+++ araig_calculators/scripts/folder_bagger_node
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+
+import sys
+import os
+from loggers.folder_bagger import FolderBagger
+import rospy
+
+if __name__ == "__main__":
+    rospy.init_node("folder_bagger")
+    bagger = FolderBagger()
