--- araig_calculators/scripts/folderlogger_node
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-import sys
-import os
-from loggers.folder_logger import FolderManagerClass
-import rospy
-
-if __name__ == "__main__":
-    rospy.init_node("folder_manager_node")
-    folder = FolderManagerClass()
