--- araig_calculators/scripts/rosbagger_node
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-
-import sys
-import os
-from loggers.rosbagger_class import RosbaggerClass
-import rospy
-
-if __name__ == "__main__":
-    rospy.init_node("rosbagger")
-    bagger = RosbaggerClass()
